+++ apps/api/src/app/admin/update-asset-profile.dto.ts
@@ -16,6 +16,10 @@ export class UpdateAssetProfileDto {
 
   @IsString()
   @IsOptional()
+  currency?: string;
+
+  @IsString()
+  @IsOptional()
   name?: string;
 
   @IsObject()