+++ apps/client/src/app/services/admin.service.ts
@@ -206,6 +206,7 @@ export class AdminService {
     assetClass,
     assetSubClass,
     comment,
+    currency,
     dataSource,
     name,
     scraperConfiguration,
@@ -218,6 +219,7 @@ export class AdminService {
         assetClass,
         assetSubClass,
         comment,
+        currency,
         name,
         scraperConfiguration,
         symbolMapping